+++ src-docs/docs/changelog.md
@@ -31,6 +31,7 @@ This is the list of what has changed since the first public version of *DuFFmpeg
 - Restricted the ability to change the color range to YUV pixel formats only.
 - Improved settings UI (moved cache options to their own section, hide After Effects on Linux...)
 - Improved codec profile management for h.264 / AVC.
+- Improved pixel format management.
 - Improved custom parameters management.
 - Code refactoring and improvements to improve performance and stability, and prepare the development of the render queue.